Responsive idea section and improved overall UI
