+++ frontend/src/App.jsx
@@ -7,7 +7,7 @@ import ProjectDetail from "./components/ProjectDetail"
 import PostProject from "./components/PostForm"
 import {BrowserRouter,Routes,Route} from "react-router-dom"
 import IdeaDetail from "./components/IdeaDetail"
-import Showcase from "./components/Showcase"
+import Showcase from "./components/Project"
 
 function App() {
